Make checks for wiki existence during backup go to right server
Previously when a server backing up from another server checks to see if
a wiki exists on the remote server, it was actually checking to see if
the wiki existed locally. This made no sesnse and caused errors when a
wiki did not exist on the remote server but a backup file was generated
anyway.